Add signout as server action, move server calls
